requirements.txt
django
google-generativeai
retry